requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
pydantic-settings==2.8.1
pika==1.3.2
python-multipart==0.0.20
